backend/requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
coreschema==0.0.4
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.2
Pillow==6.2.1
pipfile==0.0.2
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
sqlparse==0.3.0
toml==0.10.0
uritemplate==3.0.0
urllib3==1.25.6
